+++ packages/client/CHANGELOG.md
@@ -1,3 +1,5 @@
 # 0.0.17-beta
 - Add get loading for `usersClient` using `CombineLoadings` from `ht-data`
-- Remove "Show all" filter tag till clear data on change is implemented
+- Add `AllowedQueryMaps` and remove `AllowedQueryKeys`
+- Clear data of 'listAll' when `show_all` is removed
+- Add `getBaseQuery` to entities